+++ lilith/src/test/groovy/de/huxhorn/lilith/swing/actions/FocusMDCActionSpec.groovy
@@ -45,4 +45,9 @@ class FocusMDCActionSpec extends AbstractFilterActionSpecBase {
 	Class expectedConditionClass() {
 		return MDCContainsCondition.class
 	}
+
+	@Override
+	boolean isExpectingAlternativeBehavior() {
+		true
+	}
 }